requirements.txt
aiohttp
mlb-statsapi
